+++ src/TableCaption.php
@@ -14,27 +14,17 @@ namespace Webuni\CommonMark\TableExtension;
 
 use League\CommonMark\Block\Element\AbstractBlock;
 use League\CommonMark\Block\Element\InlineContainer;
-use League\CommonMark\ContextInterface;
 use League\CommonMark\Cursor;
 
-class TableCell extends AbstractBlock implements InlineContainer
+class TableCaption extends AbstractBlock implements InlineContainer
 {
-    const TYPE_HEAD = 'th';
-    const TYPE_BODY = 'td';
+    public $id;
 
-    const ALIGN_LEFT = 'left';
-    const ALIGN_RIGHT = 'right';
-    const ALIGN_CENTER = 'center';
-
-    public $type = self::TYPE_BODY;
-    public $align;
-
-    public function __construct($string = '', $type = self::TYPE_BODY, $align = null)
+    public function __construct($caption, $id = null)
     {
         parent::__construct();
-        $this->finalStringContents = $string;
-        $this->type = $type;
-        $this->align = $align;
+        $this->finalStringContents = $caption;
+        $this->id = $id;
     }
 
     public function canContain(AbstractBlock $block)
@@ -55,9 +45,5 @@ class TableCell extends AbstractBlock implements InlineContainer
     public function matchesNextLine(Cursor $cursor)
     {
         return false;
-    }
-
-    public function handleRemainingContents(ContextInterface $context, Cursor $cursor)
-    {
     }
 }