+++ src/TableParser.php
@@ -4,6 +4,7 @@
  * This is part of the webuni/commonmark-table-extension package.
  *
  * (c) Martin Hasoň <martin.hason@gmail.com>
+ * (c) Webuni s.r.o. <info@webuni.cz>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -21,6 +22,7 @@ class TableParser extends AbstractBlockParser
 {
     const REGEXP_DEFINITION = '/(?: *(:?) *-+ *(:?) *)+(?=\||$)/';
     const REGEXP_CELLS = '/(?:`[^`]*`|\\\\\\\\|\\\\\||[^|`\\\\]+)+(?=\||$)/';
+    const REGEXP_CAPTION = '/^\[(.+?)\](?:\[(.+)\])?\s*$/';
 
     public function parse(ContextInterface $context, Cursor $cursor)
     {
@@ -41,14 +43,24 @@ class TableParser extends AbstractBlockParser
         }
 
         $columns = $this->parseColumns($match);
-        $row = $this->parseRow(trim(array_pop($lines)), $columns, TableCell::TYPE_HEAD);
-        if (null === $row) {
+        $head = $this->parseRow(trim(array_pop($lines)), $columns, TableCell::TYPE_HEAD);
+        if (null === $head) {
             return false;
         }
 
         $table = new Table(function (Cursor $cursor) use (&$table, $columns) {
             $row = $this->parseRow($cursor->getLine(), $columns);
             if (null === $row) {
+                if (null !== $table->getCaption()) {
+                    return false;
+                }
+
+                if (null !== ($caption = $this->parseCaption($cursor->getLine()))) {
+                    $table->setCaption($caption);
+
+                    return true;
+                }
+
                 return false;
             }
 
@@ -57,7 +69,7 @@ class TableParser extends AbstractBlockParser
             return true;
         });
 
-        $table->getHead()->appendChild($row);
+        $table->getHead()->appendChild($head);
 
         if (count($lines) >= 1) {
             $paragraph = new Paragraph();
@@ -95,12 +107,13 @@ class TableParser extends AbstractBlockParser
     private function parseRow($line, array $columns, $type = TableCell::TYPE_BODY)
     {
         $cells = RegexHelper::matchAll(self::REGEXP_CELLS, $line);
-        if (null === $cells || !is_array($cells[0])) {
+
+        if (null === $cells || $line === $cells[0]) {
             return;
         }
 
         $row = new TableRow();
-        foreach ($cells[0] as $i => $cell) {
+        foreach ((array) $cells[0] as $i => $cell) {
             $row->appendChild(new TableCell(trim($cell), $type, isset($columns[$i]) ? $columns[$i] : null));
         }
 
@@ -109,5 +122,16 @@ class TableParser extends AbstractBlockParser
         }
 
         return $row;
+    }
+
+    private function parseCaption($line)
+    {
+        $caption = RegexHelper::matchAll(self::REGEXP_CAPTION, $line);
+
+        if (null === $caption) {
+            return;
+        }
+
+        return new TableCaption($caption[1], $caption[2]);
     }
 }