+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 CHANGELOG
 =========
 
+0.6.1 (2017-01-11)
+------------------
+
+ * Fixed parsing of one column tables
+
+0.6.0 (2016-09-26)
+------------------
+
+ * Updated to the commonmark 0.15 API
+ * Moved twig template to [webuni/commonmark-twig-renderer](https://packagist.org/packages/webuni/commonmark-twig-renderer)
+
+0.5.0 (2016-07-13)
+------------------
+
+ * Added support for table caption (MultiMarkdown)
+ * Added a template for twig renderer
+ * Updated to the commonmark 0.14 API
+
 0.4.0 (2015-09-21)
 ------------------